module with basic functions for cfpq added
