+++ project/cf_graph_recognizer.py
@@ -1,0 +1,100 @@
+"""Context-free recognizers for graphs"""
+
+from typing import Set, Tuple
+import networkx as nx
+from pyformlang.cfg import CFG
+from scipy import sparse
+
+from project import cfg_to_wcnf
+
+__all__ = ["hellings", "matrix_based"]
+
+
+def hellings(graph: nx.MultiDiGraph, cfg: CFG) -> Set[Tuple[int, str, int]]:
+    """Context-free recognizers for graph based on Hellings Algorithm"""
+    wcnf = cfg_to_wcnf(cfg)
+
+    eps_prod_heads = [p.head.value for p in wcnf.productions if not p.body]
+    term_productions = {p for p in wcnf.productions if len(p.body) == 1}
+    var_productions = {p for p in wcnf.productions if len(p.body) == 2}
+
+    r = {(v, h, v) for v in range(graph.number_of_nodes()) for h in eps_prod_heads} | {
+        (u, p.head.value, v)
+        for u, v, edge_data in graph.edges(data=True)
+        for p in term_productions
+        if p.body[0].value == edge_data["label"]
+    }
+
+    new = r.copy()
+    while new:
+        n, N, m = new.pop()
+        r_temp = set()
+
+        for u, M, v in r:
+            if v == n:
+                triplets = {
+                    (u, p.head.value, m)
+                    for p in var_productions
+                    if p.body[0].value == M
+                    and p.body[1].value == N
+                    and (u, p.head.value, m) not in r
+                }
+                r_temp |= triplets
+        r |= r_temp
+        new |= r_temp
+        r_temp.clear()
+
+        for u, M, v in r:
+            if u == m:
+                triplets = {
+                    (n, p.head.value, v)
+                    for p in var_productions
+                    if p.body[0].value == N
+                    and p.body[1].value == M
+                    and (n, p.head.value, v) not in r
+                }
+                r_temp |= triplets
+        r |= r_temp
+        new |= r_temp
+
+    return r
+
+
+def matrix_based(graph: nx.MultiDiGraph, cfg: CFG) -> Set[Tuple[int, str, int]]:
+    """Context-free recognizers for graph based on matrix multiplication"""
+    wcnf = cfg_to_wcnf(cfg)
+
+    eps_prod_heads = [p.head.value for p in wcnf.productions if not p.body]
+    term_productions = {p for p in wcnf.productions if len(p.body) == 1}
+    var_productions = {p for p in wcnf.productions if len(p.body) == 2}
+    nodes_num = graph.number_of_nodes()
+    matrices = {
+        v.value: sparse.dok_matrix((nodes_num, nodes_num), dtype=bool)
+        for v in wcnf.variables
+    }
+
+    for i, j, data in graph.edges(data=True):
+        l = data["label"]
+        for v in {p.head.value for p in term_productions if p.body[0].value == l}:
+            matrices[v][i, j] = True
+
+    for i in range(nodes_num):
+        for v in eps_prod_heads:
+            matrices[v][i, i] = True
+
+    any_changing = True
+    while any_changing:
+        any_changing = False
+        for p in var_productions:
+            old_nnz = matrices[p.head.value].nnz
+            matrices[p.head.value] += (
+                matrices[p.body[0].value] @ matrices[p.body[1].value]
+            )
+            new_nnz = matrices[p.head.value].nnz
+            any_changing = any_changing or old_nnz != new_nnz
+
+    return {
+        (u, variable, v)
+        for variable, matrix in matrices.items()
+        for u, v in zip(*matrix.nonzero())
+    }
